experiment: implement the basic planner
- Based on Gorkem's algorithm of critical nodes notion.
- Primary distinction is that we now do a secondary check for ancestors,
  since for the parallelization we want only the nodes that are in the same branch.
- Each found chunk is reduced to a subgraph on the planner, which gets propagated into
  a single task in the queue construction stage.
