+++ src/fal_planner/plan.py
@@ -1,153 +1,81 @@
-from __future__ import annotations
-
-import time
-import uuid
-from collections import defaultdict
-from dataclasses import dataclass, field
-from enum import IntEnum
-from functools import cached_property
-from typing import Any, Iterator
+import itertools
+from typing import Iterator
 
 import networkx as nx
 
-# For testing purposes, each task takes 0.2 seconds
-TIMEOUT = 0.2
-DBT_OVERHEAD = 0.05
-
-SUCCESS = 0
-FAILURE = 1
+from fal_planner.static_graph_2 import graph
 
 
-class Task:
-    def execute(self) -> int:
-        raise NotImplementedError
+def _find_subgraphs(graph: nx.DiGraph) -> Iterator[list[str]]:
+    # Initially sort the graph, and split it
+    # into chunks by using the critical node
+    # approach (post-hooks are also a thing
+    # now!!).
+
+    # Then for each chunk, check whether they have the
+    # same set of ancestors. This is important because
+    # we want to group only the things that are in
+    # the same branch.
+
+    current_stack = []
+    allowed_ancestors = set()
+
+    def split() -> Iterator[list[str]]:
+        if len(current_stack) > 1:
+            yield current_stack.copy()
+        current_stack.clear()
+        allowed_ancestors.clear()
+
+    for node in nx.topological_sort(graph):
+        properties = graph.nodes[node]
+        if properties["kind"] == "python model":
+            yield from split()
+            continue
+
+        ancestors = nx.ancestors(graph, node)
+        if not current_stack:
+            allowed_ancestors = ancestors
+
+        if not ancestors.issubset(allowed_ancestors):
+            yield from split()
+
+        current_stack.append(node)
+        allowed_ancestors |= {node, *ancestors}
+
+        if properties.get("post-hook"):
+            yield from split()
+
+    yield from split()
 
 
-@dataclass
-class DBTTask(Task):
-    selectors: list[str]
+def _reduce_subgraph(
+    graph: nx.DiGraph,
+    nodes: list[str],
+) -> None:
+    subgraph = graph.subgraph(nodes).copy()
 
-    def execute(self) -> int:
-        print(f"$ dbt run --select {' '.join(self.selectors)}")
-        time.sleep(DBT_OVERHEAD + TIMEOUT * len(self.selectors))
-        print("(DONE) $ dbt run --select {}".format(" ".join(self.selectors)))
-        return SUCCESS
+    # Use the same set of properties as the last
+    # node, since only it can have post hooks.
+    graph.add_node(subgraph, **graph.nodes[nodes[-1]].copy())
 
+    for node in nodes:
+        for predecessor in graph.predecessors(node):
+            graph.add_edge(predecessor, subgraph)
 
-@dataclass
-class FalModelTask(Task):
-    model_name: str
+        for successor in graph.successors(node):
+            graph.add_edge(subgraph, successor)
 
-    def execute(self) -> int:
-        print(f"$ fal run {self.model_name}")
-        time.sleep(TIMEOUT)
-        print(f"(DONE) $ fal run {self.model_name}")
-        return SUCCESS
+        graph.remove_node(node)
 
 
-@dataclass
-class FalHookTask(Task):
-    hook_name: str
+def plan_graph(graph: nx.DiGraph) -> nx.DiGraph:
+    # Implementation of Gorkem's Algorithm
+    new_graph = graph.copy()
 
-    def execute(self) -> int:
-        print(f"$ ./scripts/{self.hook_name}.py")
-        time.sleep(TIMEOUT)
-        print(f"(DONE) $ ./scripts/{self.hook_name}.py")
-        return SUCCESS
+    for nodes in _find_subgraphs(new_graph):
+        _reduce_subgraph(new_graph, nodes)
 
-
-@dataclass
-class Node:
-    task: Task
-    post_hooks: list[Task]
-    dependencies: list[Node] = field(default_factory=list)
-
-    @cached_property
-    def id(self) -> str:
-        return str(uuid.uuid4())
+    return new_graph
 
 
-def create_node(name: str, properties: dict) -> Node:
-    if properties["kind"] == "dbt model":
-        task = DBTTask(selectors=[name])
-    else:
-        assert properties["kind"] == "python model"
-        task = FalModelTask(model_name=name)
-
-    post_hooks = [
-        FalHookTask(hook_name=hook_name)
-        for hook_name in properties.get("post-hook", [])
-    ]
-    return Node(task, post_hooks=post_hooks)
-
-
-@dataclass
-class NodeQueue:
-    nodes: list[Node]
-    _staged_nodes: list[Node] = field(default_factory=list)
-
-    def _calculate_node_score(self, target_node: Node) -> tuple[int, int]:
-        # Determine the priority of the node by doing a bunch
-        # of calculations. This doesn't really need to be 100% precise,
-        # since if we don't have this we'll have to schedule randomly.
-
-        # 1. Number of nodes which are only waiting this node (direct dependants)
-        # 2. Number of nodes which are waiting this node and other nodes (indirect dependants)
-        # ...
-
-        direct_dependants = 0
-        indirect_dependants = 1
-
-        for node in self.nodes:
-            if node is target_node:
-                continue
-
-            if any(dependency is target_node for dependency in node.dependencies):
-                indirect_dependants += 1
-                if len(node.dependencies) == 1:
-                    direct_dependants += 1
-
-        return (direct_dependants, indirect_dependants)
-
-    def _stage_node(self, target_node: Node) -> None:
-        # When we start running a node, we don't want to simply remove it
-        # just yet (since that would unblock all of its dependencies). So
-        # we temporarily add it to a separate group of nodes (staged).
-
-        self.nodes.remove(target_node)
-        self._staged_nodes.append(target_node)
-
-    def finish(self, target_node: Node, status: int) -> None:
-        # When a staged node's execution is finished, we'll remove it
-        # alltogether and unblock all of its dependencies.
-        self._staged_nodes.remove(target_node)
-
-        assert status == SUCCESS
-        for node in self.nodes:
-            if target_node in node.dependencies:
-                node.dependencies.remove(target_node)
-
-    def iter_available_nodes(self) -> Iterator[Node]:
-        # Find all unblocked nodes (nodes without any dependencies) and
-        # use the node score algorithm to determine the priority of each
-        # node.
-        unblocked_nodes = [node for node in self.nodes if not node.dependencies]
-        unblocked_nodes.sort(key=self._calculate_node_score, reverse=True)
-
-        for node in unblocked_nodes:
-            if self.can_plan_further(node):
-                pass
-            self._stage_node(node)
-            yield node
-
-
-def load_graph(graph: nx.DiGraph) -> NodeQueue:
-    tasks = {
-        name: create_node(name, properties)
-        for name, properties in graph.nodes(data=True)
-    }
-
-    for name, task in tasks.items():
-        task.dependencies = [tasks[ancestor] for ancestor in nx.ancestors(graph, name)]
-
-    return NodeQueue(list(tasks.values()))
+new_graph = plan_graph(graph)
